requirements.txt
vncdotool>=1.2.0
Pillow>=10.0.0
paramiko>=3.0.0
